Added Custom Hook and Learned about useRef
