+++ src/useMovies.jsx
@@ -1,0 +1,53 @@
+import { useState, useEffect } from "react";
+const API_KEY = "5c9494c7";
+
+export function useMovies(query) {
+  const [movies, setMovies] = useState([]);
+  const [isLoading, setIsLoading] = useState(false);
+  const [error, setError] = useState("");
+  useEffect(
+    function () {
+      const controller = new AbortController();
+      async function fetchMovies() {
+        try {
+          setIsLoading(true);
+          setError("");
+          const res = await fetch(
+            `http://www.omdbapi.com/?apikey=${API_KEY}&s=${query}`,
+            { signal: controller.signal }
+          );
+          if (!res.ok)
+            throw new Error("Something went wrong while fetching movies");
+
+          const data = await res.json();
+
+          if (data.Response === "False") throw new Error("Movie Not found");
+
+          setMovies(data.Search);
+          // console.log(data.Search);
+          setError("");
+        } catch (err) {
+          // console.error(err.message);
+          if (err.name !== "AbortError") {
+            setError(err.message);
+          }
+        } finally {
+          setIsLoading(false);
+        }
+      }
+      if (query.length < 3) {
+        setMovies([]);
+        setError("");
+        return;
+      }
+      // handleClose();
+      fetchMovies();
+      return function () {
+        controller.abort();
+      };
+    },
+    [query]
+  );
+
+  return { movies, isLoading, error };
+}
